+++ src/main/java/todayTix/TestMain.java
@@ -5,7 +5,7 @@ import todayTix.adapters.AvailableTickets;
 public class TestMain {
 
     public static void main(String[] args) {
-        new AvailableTickets().ticketsExist();
+      //  new AvailableTickets().ticketsExist("2", "20220416", "20220420");
         System.out.println("++++++++++++++++++Ejecucion completa+++++++++++++++++++++++++");
     }
 }