+++ src/main/java/todayTix/definitions/Definition.java
@@ -1,0 +1,19 @@
+package todayTix.definitions;
+
+
+import todayTix.logs.Log;
+
+import static todayTix.logs.Log.writeOperationExeption;
+
+public abstract class Definition {
+
+    protected void execute(String name, Runnable method) {
+        Log.writeOperationStarted(name);
+        try {
+            method.run();
+            Log.writeOperationCompleted();
+        } catch (Exception e) {
+            writeOperationExeption(e);
+        }
+    }
+}