+++ src/main/java/todayTix/utils/managers/FileReaderManager.java
@@ -1,0 +1,15 @@
+package todayTix.utils.managers;
+
+import lombok.SneakyThrows;
+
+import java.nio.file.Files;
+import java.nio.file.Path;
+
+public class FileReaderManager {
+
+    @SneakyThrows
+    public String getHtmlReport(String path) {
+        return Files.readString(Path.of(path));
+    }
+
+}